requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backcall==0.2.0
bytesbufio==1.0.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comtypes==1.4.8
cryptography==43.0.3
distro==1.9.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
docopt==0.6.2
fastjsonschema==2.21.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
lxml==5.3.0
multidict==6.1.0
numpy==2.1.2
openai==0.28.0
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pickleshare==0.7.5
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.12
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rpds-py==0.22.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
XlsxWriter==3.2.0
yarl==1.16.0
